requirements.txt
plotly
speedtest-cli